refactor: http request, response의 패키지 구조 변경
