+++ study/src/test/java/thread/stage1/ConcurrencyTest.java
@@ -1,5 +1,6 @@
 package thread.stage1;
 
+import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
