+++ tomcat/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.controller;
 
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.message.request.HttpRequest;
 
 public interface Controller {
