+++ tomcat/src/main/java/nextstep/jwp/controller/rest/RegisterController.java
@@ -7,9 +7,9 @@ import nextstep.jwp.controller.Controller;
 import nextstep.jwp.controller.ResponseEntity;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.HttpMethod;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpStatusCode;
+import org.apache.coyote.http11.message.HttpMethod;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.HttpStatusCode;
 
 public class RegisterController implements Controller {
