+++ tomcat/src/main/java/nextstep/servlet/filter/Interceptor.java
@@ -1,7 +1,7 @@
 package nextstep.servlet.filter;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 
 public interface Interceptor {
