+++ tomcat/src/main/java/org/apache/catalina/servlet/Servlet.java
@@ -1,8 +1,8 @@
 package org.apache.catalina.servlet;
 
 import java.io.IOException;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 
 public interface Servlet {
