+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -8,6 +8,8 @@ import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.catalina.servlet.ServletManger;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
