+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/QueryStrings.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message.request;
 
 import java.util.Arrays;
 import java.util.Collections;
